64bit: Fix compiler warnings in regard to 64bit
vty_interface_layer3.c:584:4: warning: format '%d' expects argument of type 'int', but argument 3 has type 'long unsigned int' [-Wformat=]
    sizeof(subscr->extension)-1, VTY_NEWLINE);

+++ openbsc/src/osmo-bsc_mgcp/mgcp_main.c
@@ -156,7 +156,7 @@ static int read_call_agent(struct osmo_fd *fd, unsigned int what)
 		return -1;
 	} else if (slen > sizeof(addr)) {
 		fprintf(stderr, "Gateway received message from outerspace: %zu %zu\n",
-			slen, sizeof(addr));
+			(size_t) slen, sizeof(addr));
 		return -1;
 	}
